requirements.txt
torch==2.2.1
Pillow==10.2.0
torchvision==0.17.1
pandas==2.0.3
numpy==1.24.4
